refactor: Migrate stats calculation and pagetab copying to SubmissionPostProcessingService
